+++ packages/local_auth/local_auth_darwin/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.2
+
+* Improves compatibility with UIScene.
+* Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
+
 ## 2.0.1
 
 * Updates to Pigeon 26.